some jvmbridge cleanup; jdk 1.3 support is no longer needed, all implementations are protected, some methods removed because of simplification
